fixed some more bugs, moved layout_log into log.ch, restored focus_parent
