+++ include/log.h
@@ -1,6 +1,7 @@
 #ifndef _SWAY_LOG_H
 #define _SWAY_LOG_H
 #include <stdbool.h>
+#include "container.h"
 
 typedef enum {
 	L_SILENT = 0,
@@ -15,4 +16,5 @@ void sway_log(int verbosity, const char* format, ...) __attribute__((format(prin
 void sway_abort(const char* format, ...) __attribute__((format(printf,1,2)));
 bool sway_assert(bool condition, const char* format, ...) __attribute__((format(printf,2,3)));
 
+void layout_log(const swayc_t *c, int depth);
 #endif
