+++ sway/log.c
@@ -82,3 +82,62 @@ bool sway_assert(bool condition, const char* format, ...) {
 
 	return false;
 }
+
+#include "workspace.h"
+
+/* XXX:DEBUG:XXX */
+static void container_log(const swayc_t *c) {
+	fprintf(stderr, "focus:%c|",
+			c->is_focused ? 'F' : //Focused
+			c == active_workspace ? 'W' : //active workspace
+			c == &root_container  ? 'R' : //root
+			'X');//not any others
+	fprintf(stderr,"(%p)",c);
+	fprintf(stderr,"(p:%p)",c->parent);
+	fprintf(stderr,"(f:%p)",c->focused);
+	fprintf(stderr,"(h:%ld)",c->handle);
+	fprintf(stderr,"Type:");
+	fprintf(stderr,
+			c->type == C_ROOT   ? "Root|" :
+			c->type == C_OUTPUT ? "Output|" :
+			c->type == C_WORKSPACE ? "Workspace|" :
+			c->type == C_CONTAINER ? "Container|" :
+			c->type == C_VIEW   ? "View|" : "Unknown|");
+	fprintf(stderr,"layout:");
+	fprintf(stderr,
+			c->layout == L_NONE ? "NONE|" :
+			c->layout == L_HORIZ ? "Horiz|":
+			c->layout == L_VERT ? "Vert|":
+			c->layout == L_STACKED  ? "Stacked|":
+			c->layout == L_FLOATING ? "Floating|":
+			"Unknown|");
+	fprintf(stderr, "w:%d|h:%d|", c->width, c->height);
+	fprintf(stderr, "x:%d|y:%d|", c->x, c->y);
+	fprintf(stderr, "vis:%c|", c->visible?'t':'f');
+	fprintf(stderr, "wgt:%d|", c->weight);
+	fprintf(stderr, "name:%.16s|", c->name);
+	fprintf(stderr, "children:%d\n",c->children?c->children->length:0);
+}
+void layout_log(const swayc_t *c, int depth) {
+	int i, d;
+	int e = c->children ? c->children->length : 0;
+	container_log(c);
+	if (e) {
+		for (i = 0; i < e; ++i) {
+			fputc('|',stderr);
+			for (d = 0; d < depth; ++d) fputc('-', stderr);
+			layout_log(c->children->items[i], depth + 1);
+		}
+	}
+	if (c->type == C_WORKSPACE) {
+		e = c->floating?c->floating->length:0;
+		if (e) {
+			for (i = 0; i < e; ++i) {
+				fputc('|',stderr);
+				for (d = 0; d < depth; ++d) fputc('=', stderr);
+				layout_log(c->floating->items[i], depth + 1);
+			}
+		}
+	}
+}
+/* XXX:DEBUG:XXX */
